Backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
requests
python-dotenv